+++ build.gradle
@@ -40,20 +40,8 @@ subprojects {
 
 	//공통 dependency를 설정
 	dependencies {
-		testImplementation 'org.springframework.boot:spring-boot-starter-test'
-		testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-
 		compileOnly 'org.projectlombok:lombok'
 		annotationProcessor 'org.projectlombok:lombok'
-
-		testImplementation platform('org.junit:junit-bom:5.10.0')
-		testImplementation 'org.junit.jupiter:junit-jupiter'
-		testImplementation 'org.assertj:assertj-core:3.27.3'
-
-		implementation 'org.springframework.boot:spring-boot-starter-security'
-		implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-		runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-		runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
 	}
 
 	tasks.withType(Test).configureEach {
