+++ common/build.gradle
@@ -15,7 +15,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
 
-    //swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.6'
 }
