+++ common/src/main/java/com/pickcar/presentation/dto/cycle/CycleInfoConverter.java
@@ -1,4 +1,4 @@
-package com.pickcar.dto;
+package com.pickcar.presentation.dto.cycle;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JavaType;
