+++ common/src/main/java/com/pickcar/presentation/dto/cycle/GpsStatus.java
@@ -1,4 +1,4 @@
-package com.pickcar.dto;
+package com.pickcar.presentation.dto.cycle;
 
 public enum GpsStatus {
     NORMAL, ABNORMAL, NOT_INSTALLED, ON_ABNORMAL
