+++ consumer/build.gradle
@@ -12,16 +12,13 @@ repositories {
 }
 
 dependencies {
-    implementation project(":common")
-    implementation project(":database")
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-amqp'
-    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
     implementation 'mysql:mysql-connector-java:8.0.33'
+
+    implementation project(":common")
+    implementation project(":database")
 }
 
 bootJar {
