+++ consumer/src/main/java/com/pickcar/infrastructure/conveter/CycleInfoConverter.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.pickcar.dto.CycleInfoPayload;
+import com.pickcar.presentation.dto.cycle.CycleInfoPayload;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import java.io.IOException;
