+++ database/build.gradle
@@ -6,7 +6,6 @@ plugins {
 
 dependencies {
     implementation project(":common")
-    implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 }
