+++ database/src/main/java/com/pickcar/emulator/domain/Cycle.java
@@ -1,7 +1,7 @@
 package com.pickcar.emulator.domain;
 
-import com.pickcar.dto.CycleInfoConverter;
-import com.pickcar.dto.CycleInfoPayload;
+import com.pickcar.presentation.dto.cycle.CycleInfoConverter;
+import com.pickcar.presentation.dto.cycle.CycleInfoPayload;
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
