+++ domain/src/main/java/com/pickcar/auth/application/LoginService.java
@@ -4,8 +4,7 @@ import com.pickcar.auth.domain.SuperAdmin;
 import com.pickcar.auth.domain.User;
 import com.pickcar.auth.infrastructure.SuperAdminRepository;
 import com.pickcar.auth.infrastructure.UserRepository;
-import com.pickcar.jwt.JwtProvider;
-import com.pickcar.presentation.dto.response.SuccessResponse;
+import com.pickcar.auth.jwt.JwtProvider;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
