+++ domain/src/main/java/com/pickcar/auth/jwt/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.pickcar.jwt;
+package com.pickcar.auth.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
