+++ emulator/build.gradle
@@ -12,19 +12,14 @@ repositories {
 }
 
 dependencies {
-    implementation project(":common")
     implementation 'org.springframework.boot:spring-boot-starter-web'
+
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
-
-    // JAXB API: XML 데이터를 Java 객체로 변환하기 위해 필요
     implementation 'jakarta.xml.bind:jakarta.xml.bind-api'
-    implementation 'org.glassfish.jaxb:jaxb-runtime'
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
 
     implementation 'org.springframework.boot:spring-boot-starter-amqp'
+
+    implementation project(":common")
 }
 
 bootJar {
