+++ producer/build.gradle
@@ -12,14 +12,13 @@ repositories {
 }
 
 dependencies {
-    implementation project(":common")
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
+
     implementation 'org.springframework.boot:spring-boot-starter-amqp'
-    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
+
     implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
+
+    implementation project(":common")
 }
 
 bootJar {
