+++ producer/src/main/java/com/pickcar/dto/CyclePayload.java
@@ -2,6 +2,7 @@ package com.pickcar.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.pickcar.presentation.dto.cycle.CycleInfoPayload;
 import java.io.Serializable;
 import java.time.LocalDateTime;
 import java.util.List;
