+++ producer/src/main/java/com/pickcar/dto/EventPayload.java
@@ -2,6 +2,7 @@ package com.pickcar.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.pickcar.presentation.dto.cycle.GpsStatus;
 import java.io.Serializable;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
